refactor: componentize home and clean styles
