+++ src/components/PaginationControls/index.ts
@@ -1,0 +1,1 @@
+export { default } from './index.vue'
