+++ src/components/ProductFilters/styles.scss
@@ -1,0 +1,5 @@
+.form-control:focus,
+.form-select:focus {
+  border-color: #2e8b57;
+  box-shadow: 0 0 0 0.2rem rgba(46, 139, 87, 0.25);
+}
