+++ src/components/shared/CartPanel/index.vue
@@ -212,249 +212,5 @@ function handleImageError(event: Event) {
 }
 </script>
 
-<style scoped>
-.cart-panel {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  right: -400px;
-  width: 380px;
-  max-width: 90vw;
-  background: white;
-  border: 1px solid #dee2e6;
-  z-index: 1050;
-  overflow-y: auto;
-  transition: right 0.3s ease;
-}
 
-.cart-panel.show {
-  right: 0;
-}
-
-.cart-fixed {
-  right: 0;
-  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1) !important;
-}
-
-.cart-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(0, 0, 0, 0.5);
-  z-index: 1040;
-}
-
-.cart-header {
-  padding: 1rem;
-  border-bottom: 1px solid #eee;
-  background: #f8f9fa;
-  border-radius: 8px 8px 0 0;
-  position: sticky;
-  top: 0;
-  z-index: 1;
-}
-
-.empty-cart {
-  padding: 2rem 1rem;
-}
-
-.cart-content {
-  padding: 1rem;
-}
-
-.cart-items {
-  max-height: 300px;
-  overflow-y: auto;
-  margin-bottom: 1rem;
-}
-
-.cart-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.75rem 0;
-  border-bottom: 1px solid #f0f0f0;
-}
-
-.cart-item:last-child {
-  border-bottom: none;
-}
-
-.item-info {
-  display: flex;
-  align-items: center;
-  flex: 1;
-  gap: 0.75rem;
-}
-
-.item-image {
-  width: 48px;
-  height: 48px;
-  object-fit: cover;
-  border-radius: 4px;
-  border: 1px solid #eee;
-}
-
-.item-details {
-  flex: 1;
-  min-width: 0;
-}
-
-.item-name {
-  font-size: 0.875rem;
-  margin: 0;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.item-controls {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.quantity-controls {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  background: #f8f9fa;
-  border-radius: 20px;
-  padding: 0.25rem;
-}
-
-.quantity-controls .btn {
-  width: 28px;
-  height: 28px;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: none;
-  font-size: 0.75rem;
-}
-
-.quantity {
-  min-width: 20px;
-  text-align: center;
-  font-size: 0.875rem;
-  font-weight: 500;
-}
-
-.cart-summary {
-  border-top: 1px solid #eee;
-  padding-top: 1rem;
-  margin-bottom: 1rem;
-}
-
-.summary-line {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 0.5rem;
-}
-
-.summary-line.total {
-  font-size: 1.1rem;
-  border-top: 1px solid #eee;
-  padding-top: 0.5rem;
-  margin-top: 0.5rem;
-}
-
-.cart-actions {
-  margin-bottom: 1rem;
-}
-
-.cart-settings {
-  padding: 0.5rem 0;
-  border-top: 1px solid #eee;
-  margin-bottom: 1rem;
-}
-
-.recent-orders {
-  border-top: 1px solid #eee;
-  padding-top: 1rem;
-}
-
-.orders-title {
-  font-size: 0.875rem;
-  color: #6c757d;
-  margin-bottom: 0.75rem;
-}
-
-.order-item {
-  margin-bottom: 0.5rem;
-}
-
-.order-link {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.5rem;
-  background: #f8f9fa;
-  border-radius: 4px;
-  text-decoration: none;
-  color: inherit;
-  transition: background-color 0.2s;
-}
-
-.order-link:hover {
-  background: #e9ecef;
-  color: inherit;
-}
-
-.order-number {
-  font-weight: 500;
-  font-size: 0.875rem;
-}
-
-.order-items-count {
-  color: #6c757d;
-}
-
-/* Dark theme */
-[data-theme='dark'] .cart-panel {
-  background: #343a40;
-  border-color: #495057;
-  color: #f8f9fa;
-}
-
-[data-theme='dark'] .cart-header {
-  background: #495057;
-  border-color: #6c757d;
-}
-
-[data-theme='dark'] .cart-item {
-  border-color: #495057;
-}
-
-[data-theme='dark'] .quantity-controls {
-  background: #495057;
-}
-
-[data-theme='dark'] .cart-summary,
-[data-theme='dark'] .cart-settings,
-[data-theme='dark'] .recent-orders {
-  border-color: #495057;
-}
-
-[data-theme='dark'] .order-link {
-  background: #495057;
-  color: #f8f9fa;
-}
-
-[data-theme='dark'] .order-link:hover {
-  background: #6c757d;
-  color: #f8f9fa;
-}
-
-/* Responsive */
-@media (max-width: 767.98px) {
-  .cart-panel {
-    width: 100%;
-    border-radius: 0;
-  }
-}
-</style>
+<style scoped src="./styles.scss"></style>
