+++ src/components/shared/FooterBar/styles.scss
@@ -1,4 +1,5 @@
 footer {
   background: var(--bs-bg);
   color: var(--bs-text);
+  font-size: 0.9rem;
 }
