+++ src/components/shared/HeaderBar/styles.scss
@@ -1,3 +1,8 @@
+.navbar {
+  position: relative;
+}
+
 .navbar-brand {
+  font-weight: bold;
   color: var(--bs-primary);
 }
