+++ src/components/shared/ThemeToggle/styles.scss
@@ -1,3 +1,4 @@
 button {
+  min-width: 4rem;
   cursor: pointer;
 }
