+++ src/pages/Home/styles.scss
@@ -1,3 +1,4 @@
 .card {
   cursor: pointer;
 }
+
